# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Market API
  version: 1.0.0
  description: |
    Backend API cho **Smart Market – Chợ thông minh** (Zalo Mini App + CMS).

    ## Xác thực
    - **Mini app**: `POST /auth/zalo-login` với access token Zalo. Môi trường dev bật `ZALO_MOCK=true`
      cho phép token dạng `mock:<zaloId>:<Tên>` (vd `mock:zalo-buyer-1:Người Mua`).
    - **CMS**: `POST /auth/login` với username/email/số điện thoại + mật khẩu.
    - Sau khi đăng nhập, bấm **Authorize** và dán `accessToken` (không cần chữ "Bearer").

    ## Vai trò & phạm vi
    | Role | Phạm vi |
    |---|---|
    | super_admin | Toàn hệ thống |
    | province_admin | Các chợ trong tỉnh được gán (chủ yếu chỉ đọc + escalate phản ánh) |
    | market_manager | CRUD trong các chợ được gán |
    | user | Người mua; nếu `merchant_status=active` và có sạp → dùng được nhóm `/seller/*` |

    ## Quy ước
    - Envelope: `{ success, data, meta? }` hoặc `{ success: false, error: { code, message, details? } }`.
    - Phân trang: `?page=1&limit=20` (limit tối đa 100) → `meta { page, limit, total, totalPages }`.
    - Tiền VND số nguyên. Thời gian ISO 8601. Key JSON camelCase.
    - Tài khoản demo (seed): SA `admin@smartmarket.vn / Admin@12345`; MM `mm.trungtam@smartmarket.vn / Manager@123`.
servers:
  - url: http://localhost:3000/api/v1
    description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
...
    Search:
      name: search
      in: query
      schema: { type: string }
    MarketIdQuery:
      name: marketId
      in: query
      schema: { type: string, format: uuid }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: STATE_INVALID }
            message: { type: string, example: Đơn đang ở trạng thái "completed", không thể chuyển sang "confirmed" }
            details: { nullable: true }
    Image:
      type: object
      required: [url]
      properties:
        url: { type: string, format: uri }
...
    ComplaintStatus:
      type: string
      enum: [new, processing, resolved, rejected, escalated]
    ComplaintType:
      type: string
      enum: [product_quality, price_issue, food_safety, service_attitude, weighing_fraud, infrastructure, order_issue, other]
    StallStatus:
      type: string
      enum: [vacant, occupied, maintenance, reserved]
    DisplayStatus:
      type: string
      enum: [vacant, occupied, maintenance, reserved, expiring_soon, has_complaint]
      description: Trạng thái hiển thị (suy ra) — has_complaint > expiring_soon > trạng thái gốc
...
  responses:
    Ok:
      description: Thành công
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              data: {}
    Paginated:
      description: Danh sách có phân trang
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              data: { type: array, items: {} }
              meta:
                type: object
                properties:
                  page: { type: integer, example: 1 }
                  limit: { type: integer, example: 20 }
                  total: { type: integer, example: 135 }
                  totalPages: { type: integer, example: 7 }
    Err:
      description: Lỗi (400/401/403/404/409/429 — xem `error.code`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security: []
paths:
...
  /complaints:
    post:
      tags: [Complaints (App)]
      summary: Gửi phản ánh/khiếu nại (gắn chợ, sạp/sản phẩm nếu có)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [marketId, type, content]
              properties:
                marketId: { type: string, format: uuid }
                stallId: { type: string, format: uuid, nullable: true }
                productId: { type: string, format: uuid, nullable: true }
                type: { $ref: '#/components/schemas/ComplaintType' }
                content: { type: string }
                images: { type: array, items: { $ref: '#/components/schemas/Image' } }
      responses:
        '201': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
...
  /admin/stalls:
    get:
      tags: [Admin - Zones & Stalls]
      summary: Danh sách sạp (kèm displayStatus + hợp đồng hiện tại)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MarketIdQuery'
        - { name: zoneId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/StallStatus' } }
        - { name: displayStatus, in: query, schema: { $ref: '#/components/schemas/DisplayStatus' } }
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
...
  /admin/traders:
    get:
      tags: [Admin - Traders]
      summary: Danh sách tiểu thương theo scope
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MarketIdQuery'
        - { name: status, in: query, schema: { type: string, enum: [active, suspended, inactive] } }
        - { name: categoryId, in: query, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
...
  /admin/complaints:
    get:
      tags: [Admin - Complaints]
      summary: Danh sách phản ánh theo scope
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ComplaintStatus' } }
        - { name: type, in: query, schema: { $ref: '#/components/schemas/ComplaintType' } }
        - $ref: '#/components/parameters/MarketIdQuery'
        - { name: stallId, in: query, schema: { type: string, format: uuid } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
...
  /admin/reports/revenue:
    get:
      tags: [Admin - Reports]
      summary: Doanh thu theo ngày/tháng (đơn completed)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: groupBy, in: query, schema: { type: string, enum: [day, month], default: day } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - $ref: '#/components/parameters/MarketIdQuery'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/reports/stalls:
    get:
      tags: [Admin - Reports]
      summary: Tỷ lệ lấp đầy + hợp đồng sắp hết hạn theo chợ
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/MarketIdQuery' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/reports/traders:
    get:
      tags: [Admin - Reports]
      summary: Top tiểu thương doanh thu cao + bị phản ánh nhiều
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - $ref: '#/components/parameters/MarketIdQuery'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/reports/complaints:
    get:
      tags: [Admin - Reports]
      summary: Phản ánh theo loại/trạng thái
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - $ref: '#/components/parameters/MarketIdQuery'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/reports/approvals:
    get:
      tags: [Admin - Reports]
      summary: Hồ sơ tiểu thương theo trạng thái
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - $ref: '#/components/parameters/MarketIdQuery'
      responses:
        '200': { $ref: '#/components/responses/Ok' }

  # ================= ADMIN - SETTINGS & USERS =================
  /admin/settings:
    get:
      tags: [Admin - Settings & Users]
      summary: Toàn bộ cấu hình (SA)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/settings/{key}:
    get:
      tags: [Admin - Settings & Users]
      summary: Đọc 1 cấu hình theo key
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: key, in: path, required: true, schema: { type: string, example: product_units } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
    put:
      tags: [Admin - Settings & Users]
      summary: Cập nhật cấu hình (SA) — upsert theo key
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: key, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { description: 'JSON bất kỳ: number/array/object' }
                description: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...